changed type int to double

+++ src/Exercise8.java
@@ -4,19 +4,19 @@ public class Exercise8 {
     public static void main(String[] args) {
         System.out.println("Calculator");
         System.out.println("Enter first number: ");
-        int a;
-        int b;
+        double a;
+        double b;
         Scanner input = new Scanner(System.in);
         a = input.nextInt();
         System.out.println("Enter second number: ");
         b = input.nextInt();
-        int sum = a + b;
+        double sum = a + b;
         System.out.println("Sum is: " + sum);
-        int sub = a - b;
+        double sub = a - b;
         System.out.println("Sub is: " + sub);
-        int multi = a * b;
+        double multi = a * b;
         System.out.println("Multiplication is: " + multi);
-        int div = a / b;
+        double div = a / b;
         System.out.println("Division is: " + div);
